+++ lazyapp/app_test.go
@@ -1006,6 +1006,7 @@ func reloadEnvironmentForTest(t *testing.T) {
 		Addr             string `default:"127.0.0.1:3000"`
 		Port             int    `default:"0"`
 		ControlPlaneAddr string
+		LazyappMigrate   string `var:"LAZYAPP_MIGRATE"`
 	}]()
 	t.Cleanup(func() {
 		environment = oldEnvironment